chore(nlu): move model hash computation in engine
