+++ modules/nlu/src/backend/model-service.ts
@@ -1,5 +1,4 @@
 import * as sdk from 'botpress/sdk'
-import crypto from 'crypto'
 import fse, { WriteStream } from 'fs-extra'
 import _ from 'lodash'
 import path from 'path'
@@ -31,23 +30,6 @@ const MAX_MODELS_TO_KEEP = 2
 
 function makeFileName(hash: string, lang: string): string {
   return `${hash}.${lang}.model`
-}
-
-// we might want to make this language specific
-export function computeModelHash(
-  intents: sdk.NLU.IntentDefinition[],
-  entities: sdk.NLU.EntityDefinition[],
-  version: NLUVersionInfo,
-  lang: string
-): string {
-  const { nluVersion, langServerInfo } = version
-
-  const singleLangIntents = intents.map(i => ({ ...i, utterances: i.utterances[lang] }))
-
-  return crypto
-    .createHash('md5')
-    .update(JSON.stringify({ singleLangIntents, entities, nluVersion, langServerInfo }))
-    .digest('hex')
 }
 
 function serializeModel(ref: Model): string {
