+++ modules/nlu/src/backend/module-lifecycle/on-bot-mount.ts
@@ -53,7 +53,7 @@ export function getOnBotMount(state: NLUState) {
               return
             }
 
-            const hash = ModelService.computeModelHash(intentDefs, entityDefs, state, languageCode)
+            const hash = engine.computeModelHash(intentDefs, entityDefs, state, languageCode)
             await ModelService.pruneModels(ghost, languageCode)
             let model = await ModelService.getModel(ghost, hash, languageCode)
